Toy Idea: More Loading logic oriented internals.
So in most cases, a loader instance should be passed around the internals. Only 4 cases exist where that does not occur (ASM Injections, EntrypointClient and EntrypointServer, API getInstance).
FabricLoader is now abstract, Knot and LaunchWrapper provide the rest of the implementation.

The goal of this expirment is to was to determine how easily one could abstract away impl specific code so for example, a launch system using ModLauncher could be toyed with.